feat(frontend): complete architecture refactor with structure-frontend skill
- Created folder structure following skill conventions: hooks/, components/ui/, components/sections/, storage/, router/
- Added path aliases in vite.config.js (@hooks, @components, @storage, @api, etc.)
- Refactored auth store: removed module-level side effect, added persist middleware, created useAuthListener hook
- Fixed login bug: backend was returning 500 on 404 (user not found) instead of creating user
- Updated RegisterComponent: now registers with Firebase email/password + backend Prisma
- Implemented createBrowserRouter with lazy loading for code splitting
- Extracted Inicio.jsx into sections: HeroSection, HowItWorksSection, RifasSection, BenefitsSection
- Extracted useHeaderState hook from Header.jsx (513 lines → hook + presentational)
- Created API layer with axios interceptors for auth token injection and error handling
- Removed all console.log and console.error from frontend code
- Added nginx proxy configuration for Dokploy deployment
- Created client/docs/frontend-refactor-plan.md with 7-block plan and progress tracking
