+++ client/.env.example
@@ -1,0 +1,27 @@
+# TuRifa — Variables de Entorno
+# Copiar este archivo a .env y llenar los valores
+
+# ===========================================
+# API
+# ===========================================
+
+# URL del backend API
+VITE_API_BACKEND_URL=http://localhost:3000
+
+# ===========================================
+# Firebase
+# ===========================================
+# Configuración de Firebase (obtenida de Firebase Console > Project Settings > General > Your apps)
+
+VITE_APIKEY=
+VITE_AUTHDOMAIN=
+VITE_PROJECTID=
+VITE_STORAGE_BUCKET=
+VITE_MESSAGING_SENDER_ID=
+VITE_APP_ID=
+
+# ===========================================
+# Stripe (opcional, para pagos)
+# ===========================================
+
+VITE_STRIPE_PUBLIC_KEY=
