--- client/src/App.jsx
@@ -1,11 +1,0 @@
-
-
-const App = () => {
-  return (
-    <div>
-      <p>Hello</p>
-    </div>
-  )
-}
-
-export default App
