+++ client/src/api/rifa.js
@@ -1,18 +1,18 @@
-import axios from "axios";
+import api from "../utils/api";
 
 export const getRifas = (token) => 
-    axios.post(`${import.meta.env.VITE_API_BACKEND_URL}/create/rifa`, {token} );
+    api.post("/create/rifa", {token} );
 export const createRifa = (rifa) => 
-    axios.post(`${import.meta.env.VITE_API_BACKEND_URL}/create/rifa`, rifa);
+    api.post("/create/rifa", rifa);
 export const getAllRifas = () => 
-    axios.get(`${import.meta.env.VITE_API_BACKEND_URL}/get/all/rifas`);
+    api.get("/get/all/rifas");
 export const getRifaById = (rifaId) => 
-    axios.get(`${import.meta.env.VITE_API_BACKEND_URL}/get/rifa`, {params : {rifaId}});
+    api.get("/get/rifa", {params : {rifaId}});
 export const updatePartialRifa = (id, fieldsToUpdate) => 
-    axios.patch(`${import.meta.env.VITE_API_BACKEND_URL}/update/rifa`, fieldsToUpdate, {params : {id}});
+    api.patch("/update/rifa", fieldsToUpdate, {params : {id}});
 export const decrementRifaTicket = (rifaId, amount) => 
-    axios.patch(`${import.meta.env.VITE_API_BACKEND_URL}/decrement/ticket`, {rifaId, amount});
+    api.patch("/decrement/ticket", {rifaId, amount});
 export const checkRifaForDraw = (rifaId, token) => 
-    axios.post(`${import.meta.env.VITE_API_BACKEND_URL}/rifas/${rifaId}/check-draw`, { token });
+    api.post(`/rifas/${rifaId}/check-draw`, { token });
 export const realizarSorteo = (rifaId, token) => 
-    axios.post(`${import.meta.env.VITE_API_BACKEND_URL}/rifas/${rifaId}/sorteo`, { token });
+    api.post(`/rifas/${rifaId}/sorteo`, { token });
