+++ client/src/api/ticket.js
@@ -1,5 +1,5 @@
-import axios  from "axios";
+import api from "../utils/api";
 
-export const createTicket = (ticket) => axios.post(`${import.meta.env.VITE_API_BACKEND_URL}/create/ticket`, ticket);
-export const getAllTicketsUser = (userId) => axios.post(`${import.meta.env.VITE_API_BACKEND_URL}/get/all/tickets`, {userId});
-export const getTicketsByRifaId = (rifaId) => axios.get(`${import.meta.env.VITE_API_BACKEND_URL}/tickets/rifa/${rifaId}`);
+export const createTicket = (ticket) => api.post("/create/ticket", ticket);
+export const getAllTicketsUser = (userId) => api.post("/get/all/tickets", {userId});
+export const getTicketsByRifaId = (rifaId) => api.get(`/tickets/rifa/${rifaId}`);
