+++ client/src/layout/Layout.jsx
@@ -1,12 +1,17 @@
+import { Outlet } from "react-router-dom";
+import { useEffect } from "react";
+import Footer from "./footer/Footer";
+import Header from "./header/Header";
+import useAuthListener from "../hooks/useAuthListener";
 
-import Footer from "./footer/Footer"
-import Header from "./header/Header"
+const Layout = () => {
+  // Inicializar listener de auth - solo se ejecuta una vez al montar
+  useAuthListener();
 
-const Layout = ({children}) => {
   return (
   <div className="layout">
     <Header />
-    <main className="layout-content">{children}</main>
+    <main className="layout-content"><Outlet /></main>
     <Footer />
   </div> 
   )
