+++ client/src/pages/home/Home.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import useAuthStore from "../../store/auth-store/use-auth-store";
+import useAuthStore from "@storage/auth-store/use-auth-store";
 
 const Home = () => {
   const { useLooged } = useAuthStore();
